+++ src/audio/audio.h
@@ -20,7 +20,7 @@ inline SAMPLE &operator -= (SAMPLE &a, SAMPLE b) { return (SAMPLE &)(((int &)a) 
 
 // Audio channel commands
 
-enum class CMD : LONG {
+enum class CMD : int {
    START_SEQUENCE=1,
    END_SEQUENCE,
    SAMPLE,
@@ -39,7 +39,7 @@ enum class CMD : LONG {
 
 // Sample shift - value used for converting total data size down to samples.
 
-inline const LONG sample_shift(const SFM Type)
+inline const int sample_shift(const SFM Type)
 {
    switch (Type) {
       default: return 0;
@@ -60,8 +60,8 @@ typedef struct _GUID {
 typedef struct WAVEFormat {
    int16_t Format;               // Type of WAVE data in the chunk: RAW or ADPCM
    int16_t Channels;             // Number of channels, 1=mono, 2=stereo
-   LONG Frequency;            // Playback frequency
-   LONG AvgBytesPerSecond;    // Channels * SamplesPerSecond * (BitsPerSample / 8)
+   int Frequency;            // Playback frequency
+   int AvgBytesPerSecond;    // Channels * SamplesPerSecond * (BitsPerSample / 8)
    int16_t BlockAlign;           // Channels * (BitsPerSample / 8)
    int16_t BitsPerSample;        // Bits per sample
    int16_t ExtraLength;
@@ -74,14 +74,14 @@ typedef struct {
     int16_t wSamplesPerBlock;
     int16_t wReserved;
   } Samples;
-  LONG dwChannelMask;         // Set to 0x3 for the left and right speakers
+  int dwChannelMask;         // Set to 0x3 for the left and right speakers
   GUID SubFormat;
 } WAVEFORMATEXTENSIBLE;
 
-typedef LONG (*MixRoutine)(APTR, LONG, LONG, LONG, float, float, float **);
+typedef int (*MixRoutine)(APTR, int, int, int, float, float, float **);
 
 // Function to set mixing step for thread-safe operation
-void set_mix_step(LONG step);
+void set_mix_step(int step);
 
 static const int16_t WAVE_RAW   = 0x0001;  // Uncompressed waveform data.
 static const int16_t WAVE_ADPCM = 0x0002;  // ADPCM compressed waveform data.
@@ -89,7 +89,7 @@ static const int16_t WAVE_FLOAT = 0x0003;  // Uncompressed floating point wavefo
 
 static const int16_t WAVE_FORMAT_EXTENSIBLE = int16_t(0xfffe);
 
-const LONG DEFAULT_BUFFER_SIZE = 8096; // Measured in samples, not bytes
+const int DEFAULT_BUFFER_SIZE = 8096; // Measured in samples, not bytes
 
 struct PlatformData { void *Void; };
 
@@ -139,10 +139,10 @@ struct AudioSample {
 
 struct AudioCommand {
    CMD  CommandID;    // Command ID
-   LONG Handle;       // Channel handle
+   int Handle;       // Channel handle
    std::variant<double,int,bool> Data; // Special data related to the command ID
 
-   AudioCommand(CMD pCommandID, LONG pHandle, double pData = 0) :
+   AudioCommand(CMD pCommandID, int pHandle, double pData = 0) :
       CommandID(pCommandID), Handle(pHandle), Data(pData) { }
 
    AudioCommand() = default;
@@ -156,11 +156,11 @@ struct AudioChannel {
    double   Volume;         // Playing volume (0 - 1.0)
    double   Pan;            // Pan value (-1.0 - 1.0)
    int64_t    EndTime;        // Anticipated end-time of playing the current sample, if OnStop is defined in the sample.
-   LONG     SampleHandle;   // Sample index, direct lookup into extAudio->Samples
+   int     SampleHandle;   // Sample index, direct lookup into extAudio->Samples
    CHF      Flags;          // Special flags
-   LONG     Position;       // Current playing/mixing byte position within Sample.
-   LONG     Frequency;      // Playback frequency
-   LONG     PositionLow;    // Playing position, lower bits
+   int     Position;       // Current playing/mixing byte position within Sample.
+   int     Frequency;      // Playback frequency
+   int     PositionLow;    // Playing position, lower bits
    int8_t     Priority;       // Priority of the sound that has been assigned to this channel
    CHS      State;          // Channel state
    int8_t     LoopIndex;      // The current active loop (either 0, 1 or 2)
@@ -179,7 +179,7 @@ struct ChannelSet {
    std::vector<AudioChannel> Channel;  // Array of channel objects
    std::vector<AudioChannel> Shadow;   // Array of shadow channels for oversampling
    std::vector<AudioCommand> Commands; // Buffered commands.
-   LONG UpdateRate;   // Update rate, measured in milliseconds
+   int UpdateRate;   // Update rate, measured in milliseconds
    SAMPLE MixLeft;    // Amount of mix elements left before the next command-update occurs
 
    ChannelSet() {
@@ -217,8 +217,8 @@ struct VolumeCtl {
 
 struct MixTimer {
    int64_t Time;
-   LONG  SampleHandle;
-   MixTimer(int64_t pTime, LONG pHandle) : Time(pTime), SampleHandle(pHandle) { }
+   int  SampleHandle;
+   MixTimer(int64_t pTime, int pHandle) : Time(pTime), SampleHandle(pHandle) { }
 };
 
 class extAudio : public objAudio {
@@ -245,22 +245,22 @@ class extAudio : public objAudio {
    TIMER   Timer;
    BYTELEN MixBufferSize;
    SAMPLE  MixElements;
-   LONG    MaxChannels;    // Recommended maximum mixing channels for Sound class
+   int    MaxChannels;    // Recommended maximum mixing channels for Sound class
    std::string Device;
    int8_t  DriverBitSize;  // Target sample bit size; accounts for stereo channel
    bool  Stereo;
    bool  Mute;
    bool  Initialising;
 
-   inline struct AudioChannel * GetChannel(LONG Handle) {
+   inline struct AudioChannel * GetChannel(int Handle) {
       return &this->Sets[Handle>>16].Channel[Handle & 0xffff];
    }
 
-   inline struct AudioChannel * GetShadow(LONG Handle) {
+   inline struct AudioChannel * GetShadow(int Handle) {
       return &this->Sets[Handle>>16].Shadow[Handle & 0xffff];
    }
 
-   inline SAMPLE MixLeft(LONG Value) {
+   inline SAMPLE MixLeft(int Value) {
       if (!Value) return SAMPLE(0);
       return SAMPLE((((100 * (int64_t)OutputRate) / (Value * 40)) + 1) & 0xfffffffe);
    }
@@ -287,9 +287,9 @@ class extSound : public objSound {
    STRING Path;
    TIMER  StreamTimer;        // Timer to regularly trigger for provisioning streaming data.
    TIMER  PlaybackTimer;      // Timer to trigger when playback ends.
-   LONG   Format;             // The format of the sound data
-   LONG   DataOffset;         // Start of raw audio data within the source file
-   LONG   Note;               // Note to play back (e.g. C, C#, G...)
+   int   Format;             // The format of the sound data
+   int   DataOffset;         // Start of raw audio data within the source file
+   int   Note;               // Note to play back (e.g. C, C#, G...)
    char   NoteString[4];
    bool   Active;             // True once the sound is registered with the audio driver or mixer.
 };
