+++ src/audio/class_audio.cpp
@@ -31,7 +31,7 @@ static ERR init_audio(extAudio *Self)
    Self->Stereo       = true;
    Self->MasterVolume = Self->Volumes[0].Channels[0];
    Self->Volumes[0].Flags |= VCF::MONO;
-   for (LONG i=1; i < std::ssize(Self->Volumes[0].Channels); i++) Self->Volumes[0].Channels[i] = -1;
+   for (int i=1; i < std::ssize(Self->Volumes[0].Channels); i++) Self->Volumes[0].Channels[i] = -1;
    if ((Self->Volumes[0].Flags & VCF::MUTE) != VCF::NIL) Self->Mute = true;
    else Self->Mute = false;
 
@@ -126,7 +126,7 @@ static ERR AUDIO_Activate(extAudio *Self)
 
    // Allocate a floating-point mixing buffer
 
-   const LONG mixbitsize = Self->Stereo ? sizeof(float) * 2 : sizeof(float);
+   const int mixbitsize = Self->Stereo ? sizeof(float) * 2 : sizeof(float);
 
    Self->MixBufferSize = BYTELEN((F2T((mixbitsize * Self->OutputRate) * (MIX_INTERVAL * 1.5)) + 15) & (~15));
    Self->MixElements   = SAMPLE(Self->MixBufferSize / mixbitsize);
@@ -231,7 +231,7 @@ ERR AUDIO_AddSample(extAudio *Self, struct snd::AddSample *Args)
 
    // Find an unused sample block.  If there is none, increase the size of the sample management area.
 
-   LONG idx;
+   int idx;
    for (idx=1; idx < std::ssize(Self->Samples); idx++) {
       if (!Self->Samples[idx].Data) break;
    }
@@ -333,7 +333,7 @@ AllocMemory: Failed to allocate the stream buffer.
 
 *********************************************************************************************************************/
 
-static const LONG MAX_STREAM_BUFFER = 16 * 1024; // Max stream buffer length in bytes
+static const int MAX_STREAM_BUFFER = 16 * 1024; // Max stream buffer length in bytes
 
 static ERR AUDIO_AddStream(extAudio *Self, struct snd::AddStream *Args)
 {
@@ -346,7 +346,7 @@ static ERR AUDIO_AddStream(extAudio *Self, struct snd::AddStream *Args)
 
    // Find an unused sample block.  If there is none, increase the size of the sample management area.
 
-   LONG idx;
+   int idx;
    for (idx=1; idx < std::ssize(Self->Samples); idx++) {
       if (!Self->Samples[idx].Data) break;
    }
@@ -355,7 +355,7 @@ static ERR AUDIO_AddStream(extAudio *Self, struct snd::AddStream *Args)
 
    auto shift = sample_shift(Args->SampleFormat);
 
-   LONG buffer_len;
+   int buffer_len;
    if (Args->SampleLength > 0) {
       buffer_len = Args->SampleLength / 2;
       if (buffer_len > MAX_STREAM_BUFFER) buffer_len = MAX_STREAM_BUFFER;
@@ -467,7 +467,7 @@ static ERR AUDIO_CloseChannels(extAudio *Self, struct snd::CloseChannels *Args)
 
    log.branch("Handle: $%.8x", Args->Handle);
 
-   LONG index = Args->Handle>>16;
+   int index = Args->Handle>>16;
    if ((index < 0) or (index >= std::ssize(Self->Sets))) log.warning(ERR::Args);
 
    Self->Sets[index].clear(); // We can't erase because that would mess up other channel handles.
@@ -571,15 +571,15 @@ static ERR AUDIO_NewObject(extAudio *Self)
 #ifdef __linux__
    Self->Volumes.resize(2);
    Self->Volumes[0].Name = "Master";
-   for (LONG i=0; i < std::ssize(Self->Volumes[0].Channels); i++) Self->Volumes[0].Channels[i] = 1.0;
+   for (int i=0; i < std::ssize(Self->Volumes[0].Channels); i++) Self->Volumes[0].Channels[i] = 1.0;
 
    Self->Volumes[1].Name = "PCM";
-   for (LONG i=0; i < std::ssize(Self->Volumes[1].Channels); i++) Self->Volumes[1].Channels[i] = 1.0;
+   for (int i=0; i < std::ssize(Self->Volumes[1].Channels); i++) Self->Volumes[1].Channels[i] = 1.0;
 #else
    Self->Volumes.resize(1);
    Self->Volumes[0].Name = "Master";
    Self->Volumes[0].Channels[0] = 1.0;
-   for (LONG i=1; i < std::ssize(Self->Volumes[0].Channels); i++) Self->Volumes[0].Channels[i] = -1;
+   for (int i=1; i < std::ssize(Self->Volumes[0].Channels); i++) Self->Volumes[0].Channels[i] = -1;
 #endif
 
    load_config(Self);
@@ -635,7 +635,7 @@ static ERR AUDIO_OpenChannels(extAudio *Self, struct snd::OpenChannels *Args)
 
    // Bear in mind that the +1 is for channel set 0 being a dummy entry.
 
-   LONG index = std::ssize(Self->Sets) + 1;
+   int index = std::ssize(Self->Sets) + 1;
    Self->Sets.resize(index+1);
 
    Self->Sets[index].Channel.resize(Args->Total);
@@ -731,7 +731,7 @@ static ERR AUDIO_SaveToObject(extAudio *Self, struct acSaveToObject *Args)
       else config->write("AUDIO", "Device", "default");
 
       if ((!Self->Volumes.empty()) and ((Self->Flags & ADF::SYSTEM_WIDE) != ADF::NIL)) {
-         for (LONG i=0; i < std::ssize(Self->Volumes); i++) {
+         for (int i=0; i < std::ssize(Self->Volumes); i++) {
             std::ostringstream out;
             if ((Self->Volumes[i].Flags & VCF::MUTE) != VCF::NIL) out << "1,[";
             else out << "0,[";
@@ -739,7 +739,7 @@ static ERR AUDIO_SaveToObject(extAudio *Self, struct acSaveToObject *Args)
             if ((Self->Volumes[i].Flags & VCF::MONO) != VCF::NIL) {
                out << Self->Volumes[i].Channels[0];
             }
-            else for (LONG c=0; c < std::ssize(Self->Volumes[i].Channels); c++) {
+            else for (int c=0; c < std::ssize(Self->Volumes[i].Channels); c++) {
                if (c > 0) out << ',';
                out << Self->Volumes[i].Channels[c];
             }
@@ -753,10 +753,10 @@ static ERR AUDIO_SaveToObject(extAudio *Self, struct acSaveToObject *Args)
 
    snd_mixer_selem_id_t *sid;
    snd_mixer_elem_t *elem;
-   LONG left, right;
-   LONG pmin, pmax;
-   LONG mute;
-   LONG i;
+   int left, right;
+   int pmin, pmax;
+   int mute;
+   int i;
 
    snd_mixer_selem_id_alloca(&sid);
    snd_mixer_selem_id_set_index(sid, 0);
@@ -890,7 +890,7 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
 
 #ifdef ALSA_ENABLED
 
-   LONG index;
+   int index;
    snd_mixer_selem_id_t *sid;
    snd_mixer_elem_t *elem;
    long pmin, pmax;
@@ -909,11 +909,11 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
          if (iequals(Args->Name, Self->Volumes[index].Name)) break;
       }
 
-      if (index IS (LONG)Self->Volumes.size()) return ERR::Search;
+      if (index IS (int)Self->Volumes.size()) return ERR::Search;
    }
    else {
       index = Args->Index;
-      if ((index < 0) or (index >= (LONG)Self->Volumes.size())) return ERR::OutOfRange;
+      if ((index < 0) or (index >= (int)Self->Volumes.size())) return ERR::OutOfRange;
    }
 
    if (iequals("Master", Self->Volumes[index].Name)) {
@@ -933,7 +933,7 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
 
    // Apply the volume
 
-   log.branch("%s: %.2f, Flags: $%.8x", Self->Volumes[index].Name.c_str(), Args->Volume, (LONG)Args->Flags);
+   log.branch("%s: %.2f, Flags: $%.8x", Self->Volumes[index].Name.c_str(), Args->Volume, (int)Args->Flags);
 
    snd_mixer_selem_id_alloca(&sid);
    snd_mixer_selem_id_set_index(sid,0);
@@ -953,7 +953,7 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
 
       double vol = Args->Volume;
       if (vol > 1.0) vol = 1.0;
-      LONG lvol = F2T(double(pmin) + (double(pmax - pmin) * vol));
+      int lvol = F2T(double(pmin) + (double(pmax - pmin) * vol));
 
       if ((Self->Volumes[index].Flags & VCF::CAPTURE) != VCF::NIL) {
          snd_mixer_selem_set_capture_volume_all(elem, lvol);
@@ -965,13 +965,13 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
       }
       else {
          if (Args->Channel IS -1) {
-            for (LONG channel=0; channel < (LONG)Self->Volumes[index].Channels.size(); channel++) {
+            for (int channel=0; channel < (int)Self->Volumes[index].Channels.size(); channel++) {
                if (Self->Volumes[index].Channels[channel] >= 0) {
                   Self->Volumes[index].Channels[channel] = vol;
                }
             }
          }
-         else if ((Args->Channel >= 0) and (Args->Channel < LONG(Self->Volumes[index].Channels.size()))) {
+         else if ((Args->Channel >= 0) and (Args->Channel < int(Self->Volumes[index].Channels.size()))) {
             Self->Volumes[index].Channels[Args->Channel] = vol;
          }
       }
@@ -979,12 +979,12 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
 
    if ((Args->Flags & SVF::UNMUTE) != SVF::NIL) {
       if ((snd_mixer_selem_has_capture_switch(elem)) and (!snd_mixer_selem_has_playback_switch(elem))) {
-         for (LONG chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
+         for (int chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
             snd_mixer_selem_set_capture_switch(elem, (snd_mixer_selem_channel_id_t)chn, 1);
          }
       }
       else if (snd_mixer_selem_has_playback_switch(elem)) {
-         for (LONG chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
+         for (int chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
             snd_mixer_selem_set_playback_switch(elem, (snd_mixer_selem_channel_id_t)chn, 1);
          }
       }
@@ -992,12 +992,12 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
    }
    else if ((Args->Flags & SVF::MUTE) != SVF::NIL) {
       if ((snd_mixer_selem_has_capture_switch(elem)) and (!snd_mixer_selem_has_playback_switch(elem))) {
-         for (LONG chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
+         for (int chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
             snd_mixer_selem_set_capture_switch(elem, (snd_mixer_selem_channel_id_t)chn, 0);
          }
       }
       else if (snd_mixer_selem_has_playback_switch(elem)) {
-         for (LONG chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
+         for (int chn=0; chn <= SND_MIXER_SCHN_LAST; chn++) {
             snd_mixer_selem_set_playback_switch(elem, (snd_mixer_selem_channel_id_t)chn, 0);
          }
       }
@@ -1011,7 +1011,7 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
 
 #else
 
-   LONG index;
+   int index;
 
    if (!Args) return log.warning(ERR::NullArgs);
    if (((Args->Volume < 0) or (Args->Volume > 1.0)) and (Args->Volume != -1)) {
@@ -1022,15 +1022,15 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
    // Determine what mixer we are going to adjust
 
    if (Args->Name) {
-      for (index=0; index < (LONG)Self->Volumes.size(); index++) {
+      for (index=0; index < (int)Self->Volumes.size(); index++) {
          if (iequals(Args->Name, Self->Volumes[index].Name)) break;
       }
 
-      if (index IS (LONG)Self->Volumes.size()) return ERR::Search;
+      if (index IS (int)Self->Volumes.size()) return ERR::Search;
    }
    else {
       index = Args->Index;
-      if ((index < 0) or (index >= (LONG)Self->Volumes.size())) return ERR::OutOfRange;
+      if ((index < 0) or (index >= (int)Self->Volumes.size())) return ERR::OutOfRange;
    }
 
    if (iequals("Master", Self->Volumes[index].Name)) {
@@ -1050,7 +1050,7 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
 
    // Apply the volume
 
-   log.branch("%s: %.2f, Flags: $%.8x", Self->Volumes[index].Name.c_str(), Args->Volume, (LONG)Args->Flags);
+   log.branch("%s: %.2f, Flags: $%.8x", Self->Volumes[index].Name.c_str(), Args->Volume, (int)Args->Flags);
 
    if ((Args->Volume >= 0) and (Args->Volume <= 1.0)) {
       if ((Self->Volumes[index].Flags & VCF::MONO) != VCF::NIL) {
@@ -1058,13 +1058,13 @@ static ERR AUDIO_SetVolume(extAudio *Self, struct snd::SetVolume *Args)
       }
       else {
          if (Args->Channel IS -1) {
-            for (LONG channel=0; channel < (LONG)Self->Volumes[0].Channels.size(); channel++) {
+            for (int channel=0; channel < (int)Self->Volumes[0].Channels.size(); channel++) {
                if (Self->Volumes[index].Channels[channel] >= 0) {
                   Self->Volumes[index].Channels[channel] = Args->Volume;
                }
             }
          }
-         else if ((Args->Channel >= 0) and (Args->Channel < LONG(Self->Volumes[index].Channels.size()))) {
+         else if ((Args->Channel >= 0) and (Args->Channel < int(Self->Volumes[index].Channels.size()))) {
             Self->Volumes[index].Channels[Args->Channel] = Args->Volume;
          }
       }
@@ -1092,7 +1092,7 @@ being the recommended value for CD quality playback.
 
 *********************************************************************************************************************/
 
-static ERR SET_BitDepth(extAudio *Self, LONG Value)
+static ERR SET_BitDepth(extAudio *Self, int Value)
 {
    if (Value IS 16) Self->BitDepth = 16;
    else if (Value IS 8) Self->BitDepth = 8;
@@ -1200,10 +1200,10 @@ field to `false`.  Muting does not disable the audio system, which is achieved b
 
 *********************************************************************************************************************/
 
-static ERR GET_Mute(extAudio *Self, LONG *Value)
+static ERR GET_Mute(extAudio *Self, int *Value)
 {
    *Value = FALSE;
-   for (LONG i=0; i < std::ssize(Self->Volumes); i++) {
+   for (int i=0; i < std::ssize(Self->Volumes); i++) {
       if (iequals("Master", Self->Volumes[i].Name)) {
          if ((Self->Volumes[i].Flags & VCF::MUTE) != VCF::NIL) *Value = TRUE;
          break;
@@ -1212,7 +1212,7 @@ static ERR GET_Mute(extAudio *Self, LONG *Value)
    return ERR::Okay;
 }
 
-static ERR SET_Mute(extAudio *Self, LONG Value)
+static ERR SET_Mute(extAudio *Self, int Value)
 {
    return Self->setVolume(0, "Master", Value ? SVF::MUTE : SVF::UNMUTE, -1, -1);
 }
@@ -1229,7 +1229,7 @@ The OutputRate can only be set prior to initialisation, further attempts to set 
 
 *********************************************************************************************************************/
 
-static ERR SET_OutputRate(extAudio *Self, LONG Value)
+static ERR SET_OutputRate(extAudio *Self, int Value)
 {
    if (Value < 0) return ERR::OutOfRange;
    else if (Value > 44100) Self->OutputRate = 44100;
@@ -1260,7 +1260,7 @@ audio consumption. More periods provide greater buffering security at the cost o
 
 *********************************************************************************************************************/
 
-static ERR SET_Periods(extAudio *Self, LONG Value)
+static ERR SET_Periods(extAudio *Self, int Value)
 {
    Self->Periods = Value;
    if (Self->Periods < 2) Self->Periods = 2;
@@ -1300,7 +1300,7 @@ sequential periods, providing time for processing whilst maintaining continuous 
 
 *********************************************************************************************************************/
 
-static ERR SET_PeriodSize(extAudio *Self, LONG Value)
+static ERR SET_PeriodSize(extAudio *Self, int Value)
 {
    Self->PeriodSize = Value;
    if (Self->PeriodSize < 1024) Self->PeriodSize = 1024;
@@ -1330,7 +1330,7 @@ Setting the Quality value automatically adjusts the following processing flags:
 
 *********************************************************************************************************************/
 
-static ERR SET_Quality(extAudio *Self, LONG Value)
+static ERR SET_Quality(extAudio *Self, int Value)
 {
    Self->Quality = Value;
 
@@ -1353,14 +1353,14 @@ Stereo: Set to `true` for stereo output and `false` for mono output.
 
 *********************************************************************************************************************/
 
-static ERR GET_Stereo(extAudio *Self, LONG *Value)
+static ERR GET_Stereo(extAudio *Self, int *Value)
 {
    if ((Self->Flags & ADF::STEREO) != ADF::NIL) *Value = TRUE;
    else *Value = FALSE;
    return ERR::Okay;
 }
 
-static ERR SET_Stereo(extAudio *Self, LONG Value)
+static ERR SET_Stereo(extAudio *Self, int Value)
 {
    if (Value IS TRUE) Self->Flags |= ADF::STEREO;
    else Self->Flags &= ~ADF::STEREO;
@@ -1383,7 +1383,7 @@ static void load_config(extAudio *Self)
       config->read("AUDIO", "Quality", Self->Quality);
       config->read("AUDIO", "BitDepth", Self->BitDepth);
 
-      LONG value;
+      int value;
       if (config->read("AUDIO", "Periods", value) IS ERR::Okay) SET_Periods(Self, value);
       if (config->read("AUDIO", "PeriodSize", value) IS ERR::Okay) SET_PeriodSize(Self, value);
       if (config->read("AUDIO", "Device", Self->Device) != ERR::Okay) Self->Device = "default";
@@ -1406,7 +1406,7 @@ static void load_config(extAudio *Self)
                Self->Volumes.clear();
                Self->Volumes.resize(keys.size());
 
-               LONG j = 0;
+               int j = 0;
                for (auto& [k, v] : keys) {
                   Self->Volumes[j].Name = k;
 
@@ -1415,7 +1415,7 @@ static void load_config(extAudio *Self)
                   while ((*str) and (*str != ',')) str++;
                   if (*str IS ',') str++;
 
-                  LONG channel = 0;
+                  int channel = 0;
                   if (*str IS '[') { // Read channel volumes
                      str++;
                      while ((*str) and (*str != ']')) {
@@ -1426,7 +1426,7 @@ static void load_config(extAudio *Self)
                      }
                   }
 
-                  while (channel < (LONG)Self->Volumes[j].Channels.size()) {
+                  while (channel < (int)Self->Volumes[j].Channels.size()) {
                      Self->Volumes[j].Channels[channel] = 0.75;
                      channel++;
                   }
